refactor(daemon): extract schtasks cmd-set codec helpers
